refactor(Télédéclaration): réorganise les tests par endpoint
